mesa: enable llvmpipe and pull in meta-clang
In the 1 year since I initially enabled llvmpipe for mesa in this layer,
oe-core added their own version of clang and made it easier to enable
for most devices.

Unfortunately, we do not fall under the category of most devices,
considering we allow meta-browser. Meta-browser requires meta-clang,
which provides conflicting versions of clang at runtime. This causes the
mesa GLSL JIT to fail, among other things.

As a result this patch now pulls in meta-clang as a dependency and
configures the default clang provider to be the version provided by
meta-clang.

This reverts commit 90389044489ac8813098e5d6314f568440ff64ec.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>