branding: Disable missing hardware support in the BSP layer
Currently MACHINE_FEATURES are disabled based on branding, but the
branding is not what determines if these are functional, the BSP version
does. While we do select the BSP version here, what is supported or not
by a given version is a property of the BSP layer. Disabling missing
MACHINE_FEATURES for each version should be handled there.

Signed-off-by: Andrew Davis <afd@ti.com>