opencv: Remove OpenCV customization and DISTRO_FEATURE
These customizations are leftover from when we had DSP acceleration
for OpenCV. Remove these.

Also, If you need OpenCV for a package you should just DEPEND on it in
the recipe, not check for it as a DISTRO_FEATURE. Fix that.

This then means "opencv" is no longer used as a DISTRO_FEATURE anywhere
in Arago or other lower layer, so drop it from DISTRO_FEATUREs.

Signed-off-by: Andrew Davis <afd@ti.com>