arago.conf: Remove unneeded LICENSE_FLAGS_ACCEPTED items
netperf is no longer marked as non-commercial (587fe5877 in meta-oe).

faad2 is still marked commercial and AAC is still covered under several
patents that may require royalties. Not sure why we decided to whitelist
this, the commit adding it has no message. So, what can be added without
explanation can be removed without explanation :)

gst-ffmpeg is no longer a package, remove this too.

Signed-off-by: Andrew Davis <afd@ti.com>