ffmpeg: Remove bbappend to disable x264
The x264 codec is no longer enabled by default in oe-core, remove the
extra disabling. The x264 codec is not the only codec that makes this
package license flag commercial, so removing that flag after only
disabling one codec is not correct. All potentially troublesome codecs
should be evaluated before setting LICENSE_FLAGS_ACCEPTED to approve
of the license flags.

Signed-off-by: Andrew Davis <afd@ti.com>