meta-arago: remove dependency on meta-arago-demos
Remove any explicit dependencies on meta-arago-demos. These were the
last of the "demos" in this layer and they were only being built for
legacy devices. Demos should primarily be in layers above this unless
they are used for functionality testing.

This patch does not modify this layer's dependency list because
meta-arago-demos was not previously listed as an explicit dependency,
despite default package groups requiring it.

These modifications are required as part of the effort to make this
layer less dependent on meta-qt5. Moving forward TI builds of arago will
enable and test Qt6 by default, but unlike before this will not be done
in a way that creates an explicit dependency in this layer. This should
allow customers and layers that consume or extend this layer to toggle
Qt support more easily.

A corresponding patch will be sent following this to remove
meta-arago-demos from master oe-layersetup configs and then one more
patch will be submitted to delete the now unreachable contents of
meta-arago-demos.

Signed-off-by: Randolph Sapp <rs@ti.com>