meta-arago-demos: remove this layer due to qt5 dep
Remove this layer now that the current recipes are unreachable. These
were the last of the "demos" in this layer and they were only being
built for legacy devices. Demos should primarily be in layers above this
unless they are used for functionality testing.

These modifications are required as part of the effort to make this
layer less dependent on meta-qt5. Moving forward TI builds of arago will
enable and test Qt6 by default, but unlike before this will not be done
in a way that creates an explicit dependency in this layer. This should
allow customers and layers that consume or extend this layer to toggle
Qt support more easily.

This should come after a patch to remove this layer from the
oe-layersetup configs. If you run into any build issues, be sure to
update your bblayers.conf accordingly.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>