meta-arago-*: remove all qt5 configs
Remove all the Qt5 specific configs, recipes, and package groups to
prepare for Qt6. Some of these package configs will still work, but they
are unecessary.

All of the conditional logic in meta-qt6 toggles the important flags
automatically for us right now. This behavior will only change if X11 is
enabled, which we currently disable.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>