meta-arago-extras: v4l-utils: Add recipe for v1.28.1
An older recipe of v4l-utils (v1.26.1) is present in the meta-oe layer
for scarthgap.

TI's 6.12 based kernel has backported multistream routing APIs from
upstream kernel, which requires v4l-utils to be the latest released
version (v1.28.1).

Thus we override the recipe in this layer to the latest release of
v4l-utils.

Signed-off-by: Yemike Abhilash Chandra <y-abhilashchandra@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>