meta-arago-test: dynamic layer for qt6 support
Add a dynamic layer for Qt6 example applications so we can run a quick
test and see if things are configured correctly.

This is being introduced into test as opposed to the distro layer as
there will soon be automated tests for Qt6, there are no explicit
dependencies for Qt in meta-arago-distro anymore, and the example
applications are currently only necessary for testing.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>