meta-arago-test: opencl_headers 2023.12.14 -> 2024.05.08
Version update for the opencl_headers. This version of the headers
is supported by the opencl-cts.

meta-openembedded rejected a backport of the headers recipe in master
to scarthgap due to their lts process. Where, new packages and major
package version changes are only allowed within the first 7 months
of a lts branch opening.

Signed-off-by: Antonios Christidis <a-christidis@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>