weston-init: do not require any outputs to start
We're in a bit of a bind now. With the addition of display hotplug, the
very real possibility of Weston trying to start before all displays are
correctly enumerated, and the lack of display managers available in
Yocto there is a need for Weston to start without any displays being
present.

This change will allow Weston to start so long as the seat has graphic
capabilities, regardless of whether or not a display is currently connected. As
displays are connected they will be enumerated as usual under the Weston
session.

This will likely come back to bite us. Users will need to remember to
disable Weston for headless applications to save on startup time and
resources. That being said, this is part of the weston-init package.
That alone should make it clear that this is only required behavior in
the event a display manager or greeter service is unable to detect and
wait for resources.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>