arago-prefs: Remove PREFERRED_PROVIDER for wpa-supplicant
A while back, the PREFERRED_PROVIDER settings for wpa-supplicant were
disabled due to a deprecation issue with OpenSSL 1.0 [1].  A short while
after that patch, a fix was provided for wpa-supplicant-wl18xx that
moved it to OpenSSL 1.1, but the PREFERRED_PROVIDER was never turned back on.

We have been happily using the mainline wpa-supplicant since that time,
so there seems to be no need to even maintain these overrides.  Remove
them.

[1] https://git.ti.com/cgit/arago-project/meta-arago/commit/meta-arago-distro/conf/distro/include/arago-prefs.inc?h=master-wip&id=413c96547f535529ce0705db0831b539b866efed

Signed-off-by: Ryan Eatmon <reatmon@ti.com>