systemd: remove sync-clocks service
This is already handled by the kernel with RTC_SYSTOHC and any of the
userspace time synchronization daemons we have (systemd-timesyncd).

This is handled only when an NTP sync occurs normally, so that the
system RTC is updated to the proper time outright and doesn't create
potential issues with the time sync daemon drift tracking between boots.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>