emptty: handle legacy init a litte better
I've found some odd behavior with this application when it comes to
daemonization and working with legacy init systems. Some of this was
related to sequencing regarding this script and the depedent seat
script, other parts are just related to this application. This fixes up
most of that behavior until it can be addressed upstream.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>