check_yocto_rules: Add new config file
TI is starting to run an automated checker for all incoming patches to
the mailing list.  One of the checks is to scan all of the files in the
repository looking for things that we have flagged as the "right way" of
doing recipes.  This file contains a set of exceptions that the checking
script is finding but can be skipped due to manual review.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>