bpftool: Fix for building bpftool-native
The bpftool is configured to use the DEBUG_PREFIX_MAP when building,
but this causes a problem when building bpftool-native.  Ubuntu 22.04
has gcc-11 as the default and DEBUG_PREFIX_MAP contains an option that
is only gcc-13 and beyond.

Simple fix is to remove the offending option when building bpftool-native.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>