conf: arago.conf: Set DEFAULTTUNE for armv7a to fix qtbase build issue
There is a build issue with qtbase on armv7a platforms.  It looks like
th compiler is trying to create assembly calls that are missing ARM
registers:

{standard input}: Assembler messages:
{standard input}:7280: Error: ARM register expected -- `ldrex r1,[s16]'
{standard input}:7282: Error: ARM register expected -- `strex r2,r1,[s16]'

Changing the DEFAULTUNE to not use the neon extensions seems to fix the
issue.  We are going to temporarily carry an override for the amv7a
platforms to lower the tune.  We are working to recreate this issue on
poky so that we can submit bug reports to try and get this fixed.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>