sbc: Fix for gcc 15 compile error
With the move to gcc 15, the code is now generating a compile error.

../sbc-2.1/sbc/sbc_primitives_armv6.c:284:9:
   error: too many arguments to function 'sbc_analyze_eight_armv6';
      expected 0, have 3

Simple fix is to drop the C standard down to gnu17 to avoid this new
error.

I am going to submit a better patch to oe-core, but we need this in
the interim to get our builds working again.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>