emptty: 0.13.0 -> 0.14.0
Update the current version of emptty to include useful upstream fixes
for the retry count function and permission issues with XDG_RUNTIME_DIR.

Additionally, fix the sysvinit service to use the proper inittab entry
instead of a separate init process which won't have the proper
permissions to take control of the target virtual console.

This is functionally a sync with the latest proposal to oe-core/master.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>