arago.conf: Disable new PACKAGECONFIG options for opencl-cts
A recent change in the opencl-cts package added a dependency to
freeglut, which requires the "x11" DISTRO_FEATURES to be enabled.
Subsequent testing shows additional errors related to the vulkan tests.

Short term fix, disable "opengl gles vulkan" for opencl-cts which removes
the freeglut requirement.  A longer term fix needs to be found since we
do not turn on "x11" for Arago.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>