pm-qa: Add PATCHTOOL = "git"
A recent commit in meta-openembedded removed the PATCHTOOL setting,
choosing to fallback to quilt.  But the setting seems to be required for
the patches to actually be applied correctly.

While we wait on meta-openembedded to either revert their commit, or
figure out why this is required, we can carry the setting here.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>