websocketd: Update to latest go build methodology
Move the recipe to use the most up to date go methodology for building
websocketd.  This removes the need for the network access during
compile.  Also, remove the INSANE_SKIP and instead list out the correct
package dependencies.  Finally, apply a small patch for the python
examples to use /usr/bin/python3 instead of /usr/bin/python so that
everything lines up with the python3-core recipe.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>