meta-arago/all: Rename machine from omap-a15 to am57xx
The "omap-a15" is a bit of a legacy holdover and is confusing.
This causes issues with mixed usage of 'dra7xx' and 'omap-a15'.
For example all EVM using the AM57xx SoC family include dra7x.inc
so both are always defined, then we use 'omap-a15' in several
places where we really mean just 'am57xx'. 'dra7xx' can't be
filtered for as needed like in the case of PRU.

AM57x is arguably the real superset device and DRA7x support
is being removed, so rename all uses of the combined 'omap-a15'
to 'am57xx'.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>