syslog-ng: Remove this recipe
The only thing this recipe did was add a custom config for syslog-ng. This
config was nearly identical to the one in meta-oe other than one line for
better systemd compat. Since 2018 that same change was made to meta-oe and
this file was no longer named correctly to be pulled in anyway. It has
been unused since then, remove it.

Signed-off-by: Andrew Davis <afd@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>