udev-aragoconf: remove legacy usb rules
These rules do not match the current structure of the USB DEVPATH and
are not actually fired.

They were initially cited as being help scripts, but I think users
should attempt to gracefully bring down the gadget interface on their
own instead of relying on a script that will attempt to reload it post
mode change.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>