udev-aragoconf: remove touchscreen related sections
In the originally cited commit, this section was upstreamed and could
be removed post Dora.

dd269c76 ("udev: fix touchscreen detection rule")

Also, the attempt to probe drivers for detected hardware as part of that
commit never made much sense. The default behavior of udev is to load
modules based on modalias values, making the rule redundant.

Signed-off-by: Randolph Sapp <rs@ti.com>
Signed-off-by: Ryan Eatmon <reatmon@ti.com>