FIXME: systemd: Remove new OSC prompt feature
Our testing flow is blocked because of the OSC prompt code that was
added to systemd 258.1.  This feature seeks to put a lot more extra
information on the prompt using escape sequences, but it is conflicting
with our board setup in some way.  All of the escaped sequences are
being printed at the prompt which makes our testing detection code
fail.

We need to get this fixed in the long run, but we also need to be able
to pass CICD in the meantime.

Signed-off-by: Ryan Eatmon <reatmon@ti.com>